requirements.txt
streamlit
requests
pandas
numpy
scikit-learn
nltk
matplotlib
plotly
python-dotenv
Pillow
